Move get_file_term_data and get_reconstruction into Client interface.  (#591)
This function adds get_file_term_data and get_reconstruction to the
Client interface, including adding those functions to the LocalClient.
It also rearranges some things in the RemoteClient struct to call the
methods from there. This is the first step in a larger effort to
separate out the file writer methods from the RemoteClient class in
order to allow for more thorough testing and simulation using the
LocalClient backend.